initiated the apidoc api documentation

+++ src/sql/01_attraction.sql
@@ -18,25 +18,25 @@ CREATE TABLE attractions (
 			primary key,
 	name text,
 	opening_time text,
-	price money,
+	price integer,
 	web_link text,
 	picture_link text,
-	city_id INTEGER REFERENCES cities (id)
+	city_id integer REFERENCES cities (id)
 );
 
 CREATE TABLE attraction_type_match (
     id serial not null
 		constraint attraction_type_match_pkey
 			primary key,
-	attraction_type_id INTEGER REFERENCES attraction_type (id),
-	attraction_id INTEGER REFERENCES attractions (id)
+	attraction_type_id integer REFERENCES attraction_type (id),
+	attraction_id integer REFERENCES attractions (id)
 );
 
 CREATE TABLE trip_list_attraction_match (
     id serial not null
 		constraint trip_list_attraction_match_pkey
 			primary key,
-	trip_list_id INTEGER REFERENCES trip_list (id),
-	attraction_id INTEGER REFERENCES attractions (id)
+	trip_list_id integer REFERENCES trip_list (id),
+	attraction_id integer REFERENCES attractions (id)
 );
